backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
channels==4.0.0
daphne==4.0.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
tensorflow==2.15.0
transformers==4.35.2
twilio==8.10.0